requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.20
langchain-chroma>=0.1.0
chromadb>=0.4.15
openai>=1.10.0
unstructured[local-inference]
pymupdf>=1.23.0
python-docx>=1.1.0
pandas>=2.1.0
numpy>=1.24.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pydantic>=2.5.0,<3.0.0
requests>=2.31.0 